requirements.txt
# AI-Powered Metadata Tagging System - Requirements
# Core Python Dependencies

# LangChain and LLM Integration
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Data Processing and Structure
pydantic>=2.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.0

# NLP and Text Processing
spacy>=3.7.0
textblob>=0.17.0
vaderSentiment>=3.3.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# File Processing and Encoding
chardet>=5.2.0
python-multipart>=0.0.6

# HTTP and API Libraries
requests>=2.31.0


fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0


# JSON and Regex (Built-in but listed for clarity)
# json - Built-in Python library
# re - Built-in Python library
# collections - Built-in Python library
# typing - Built-in Python library
# datetime - Built-in Python library
# time - Built-in Python library
# os - Built-in Python library
# traceback - Built-in Python library

# Development and Testing (Optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0

# Note: After installation, run these commands to download required models:
# python -m spacy download en_core_web_sm
# python -m spacy download en_core_web_md